Move structs into files in module
