+++ src/structs/mod.rs
@@ -1,0 +1,9 @@
+pub mod card;
+pub mod deck;
+pub mod enums;
+pub mod misc;
+pub mod player;
+pub mod hand;
+pub mod collection;
+pub mod pot;
+pub mod playerchips;
